backend/requirements.txt
# === Web Framework ===
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# === Database - PostgreSQL ===
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# === Database - MongoDB ===
motor==3.3.2
pymongo==4.6.1

# === Authentication ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# === Data Validation ===
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# === Email ===
aiosmtplib==3.0.1

# === Azure Blob Storage ===
azure-storage-blob==12.19.0

# === Environment ===
python-dotenv==1.0.0

# === Utilities ===
httpx==0.25.2
